requirements.txt
streamlit
python-dotenv
langchain
PyPDF2
faiss-cpu
openai
langchain-openai
pypdf2
langchain-community
huggingface-hub
sentence-transformers
einops